tig: add subcommand to browse an umpf and its context interactively
The 'tig' subcommand makes it possible to browse an umpf interactively
in tig(1) while showing the local and remote state of all topic
branches. This can give a broader overview how the topic branches have
developed before building another umpf.

For simplicity sake, also limit the range of the topics to the history
leading up from the umpf-base.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>